Divided the Timeline in its own JS module.
