+++ resources/assets/sass/_main.scss
@@ -512,7 +512,7 @@ main {
                     display: flex;
                     flex-direction: row;
                     justify-content: space-between;
-                    width: 738px;
+                    width: 740px;
 
                     .time-part {
                         font-size: .5rem;
@@ -530,7 +530,6 @@ main {
                     width: 722px;
                     max-width: 722px;
                     min-width: 722px;
-                    border-radius: 6px;
 
                     .time-slot-marker {
                         display: flex;
